Make gometa multicall-aware
This enhancement is synchronized with similar changes in forgemeta.
It is necessary to adapt to them.
See also: https://src.fedoraproject.org/rpms/redhat-rpm-config/pull-request/35

It is part of a series targeted at making packaging of multiple Go projects in a single spec file safer.
(Though that remains terrible and dangerous packaging practice, one spec per project will always work better.)

Adds the  -z <number> switch (similar to -a or -b in %setup, but without %setup's black magic quirks).

Removes the -u <url> switch (use -z, it is better).
THIS IS A BACKWARDS-INCOMPATIBLE CHANGE.