macros.go-rpm: Support %gobuildroot without global %goipath
For some complicated packages (such as Docker), there can be
multiple Go buildroots in the same package.  In this case, it would
be helpful to share the %gobuildroot code, but it currently expects
to have a global %goipath definition due to the line

    goipath="${goipath:-%{goipath}}"

which expands to a bash expression "${goipath:-%{goipath}" when
%goipath is undefined.  This is due to the shell not understanding
nested braces, so there is an extra "}" on the symlink that breaks
paths.

With this change, the undefined %goipath macro is dropped during
expansion so the shell expression is evaluated as expected.  This
allows multiple %gobuildroot uses in subshells with -i.

Signed-off-by: David Michael <dm0@redhat.com>